rename the aws s3 utility
